fix: convert annotate-screenshot from ESM to CJS for Node v25+ compatibility
Replace ESM imports with CJS require(), wrap top-level await in async
main() function. Rename .mjs to .cjs so Node resolves global ws module
without ESM loader issues.

Closes #162
